Notetaker/requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
huey==2.5.0  # Task queue (alternativa a Celery)

# OAuth
msal==1.34.0
httpx==0.25.2
pyjwt>=2.8.0  # Para decodificar ID tokens

# Config y Utils
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Storage
boto3==1.29.7  # Para S3 (opcional)

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

